Added tests for FileAccessor, SharedMemoryMap, FileAccessorFactory

+++ azure_functions_worker/shared_memory_data_transfer/file_accessor_unix.py
@@ -20,6 +20,10 @@ class FileAccessorUnix(FileAccessor):
             mem_map_name: str,
             mem_map_size: int,
             access: int = mmap.ACCESS_READ) -> Optional[mmap.mmap]:
+        if mem_map_name is None or mem_map_name == '':
+            raise Exception('Cannot open memory map. Invalid name.')
+        if mem_map_size < 0:
+            raise Exception('Cannot open memory map. Invalid size.')
         fd = self._open_mem_map_file(mem_map_name)
         if fd is None:
             return None
@@ -28,6 +32,10 @@ class FileAccessorUnix(FileAccessor):
 
     def create_mem_map(self, mem_map_name: str, mem_map_size: int) \
             -> Optional[mmap.mmap]:
+        if mem_map_name is None or mem_map_name == '':
+            raise Exception('Cannot create memory map. Invalid name.')
+        if mem_map_size <= 0:
+            raise Exception('Cannot create memory map. Invalid size.')
         fd = self._create_mem_map_file(mem_map_name, mem_map_size)
         if fd is None:
             return None
@@ -38,6 +46,8 @@ class FileAccessorUnix(FileAccessor):
         return mem_map
 
     def delete_mem_map(self, mem_map_name: str, mem_map: mmap.mmap) -> bool:
+        if mem_map_name is None or mem_map_name == '':
+            raise Exception('Cannot delete memory map. Invalid name.')
         try:
             fd = self._open_mem_map_file(mem_map_name)
             os.remove(fd.name)
